refactor: 컴포넌트 파일 구조에 맞게 폴더 구조 수정
