+++ src/pages/user/SalaryCorrectionPage/SalaryCorrectionHeader.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import { colors } from '../../../../styles';
+import { colors } from '../../../styles';
 
 type SalaryCorrectionHeaderProps = {
   isVisible: boolean;
