+++ src/pages/user/SalaryDetailsPage/SalaryDetailsMain/SalaryDetailsMain.tsx
