+++ src/pages/user/index.ts
@@ -2,5 +2,5 @@ export { default as UserHomePage } from './UserHomePage';
 export { default as SchedulePage } from './SchedulePage';
 export { default as ProfilePage } from './Profile';
 export { default as AttendancePage } from './AttendancePage';
-export { default as SalaryCorrectionPage } from './SalaryCorrectionPage';
-export { default as SalaryDetailsPage } from './SalaryDetailsPage';
+export { default as SalaryCorrectionPage } from './SalaryCorrectionPage/SalaryCorrectionPage';
+export { default as SalaryDetailsPage } from './SalaryDetailsPage/SalaryDetailsPage';
